[IMP] split composer test for parallelization
The tests of the composer takes a long long time, and most of it is spent
in await nextTick(). The main problem is that jest runs all the tests
sequentially. We can speed up the testing by splitting the composer tests
into multiple test files, so jest run them in parallel.

I went from ~55sec for npm run test to ~30sec.
